src/requirements.txt
line-bot-sdk
boto3
urllib3==1.26.6
aws-sam-cli
